src/requirements.txt
uvicorn==0.17.5
pandas==1.4.0
fastapi==0.73.0
tqdm==4.61.2
SQLAlchemy==1.4.25
python-dotenv==0.19.2
psycopg2-binary==2.9.3
dask==2021.9.0